:boom: create userLearningObjectDatastore interface
