+++ src/LearningObjectSearch/interactors/searchAllUserObjects/searchAllUserObjects.spec.ts
@@ -1,7 +1,16 @@
+import { Stubs } from '../../../tests/stubs';
+import { searchUserObjects } from './searchAllUserObjects';
+import { MockDataStore } from '../../../tests/mock-drivers/MockDataStore';
+
+const stubs = new Stubs();
+
 describe('If draftsOnly is specified', () => {
     describe('and the requester is not the author and is not a priveleged user', () => {
         it('should throw an error', () => {
+            stubs.searchUserObjectsUserToken 
+            searchUserObjects({
 
+            })
         });
     });
     describe('and the requester does not specify status', () => {
